Join the typedoc_pydantic back into typedoc
and rearrange the order of things a bit.
